requirements.txt
streamlit
pydeck
pandas
altair
matplotlib.pyplot